feat: breaking the browser use tools by types
